#1293 Moved EntityKeyExpressionGenerator to pojo-bl.
